Implement VK_KHR_synchronization2 * And some descriptor set improvements.

+++ include/garden/graphics/memory.hpp
@@ -83,8 +83,8 @@ public:
 	 */
 	struct BarrierState
 	{
-		uint32 access = 0; /**< Memory access flags. (Internal API format) */
-		uint32 stage = 0;  /**< Pipeline stages. (Internal API format) */
+		uint64 access = 0; /**< Memory access flags. (Internal API format) */
+		uint64 stage = 0;  /**< Pipeline stages. (Internal API format) */
 	};
 protected:
 	void* allocation = nullptr;
